+++ src/lib.rs
@@ -1,5 +1,8 @@
-pub mod config;
-pub mod language_name;
+mod config;
+mod language_name;
 pub mod loader;
 mod range_union;
 pub mod searches;
+
+pub use config::{Config, QueryCompiler};
+pub use language_name::LanguageName;